Moving the selectos to the plan folder
