+++ client/state/sites/plans/selectors/index.js
@@ -5,12 +5,15 @@ export {
 	getPlansBySite,
 	getPlansBySiteId,
 } from 'calypso/state/sites/plans/selectors/get-plans-by-site';
+export { default as getECommerceTrialDaysLeft } from './get-ecommerce-trial-days-left';
+export { default as getECommerceTrialExpiration } from './get-ecommerce-trial-expiration';
 export { getSitePlan } from 'calypso/state/sites/plans/selectors/get-site-plan';
 export { getSitePlanRawPrice } from 'calypso/state/sites/plans/selectors/get-site-plan-raw-price';
 export { getSitePlanSlug } from 'calypso/state/sites/plans/selectors/get-site-plan-slug';
 export { hasDomainCredit } from 'calypso/state/sites/plans/selectors/has-domain-credit';
 export { isCurrentPlanExpiring } from 'calypso/state/sites/plans/selectors/is-current-plan-expiring';
 export { isCurrentUserCurrentPlanOwner } from 'calypso/state/sites/plans/selectors/is-current-user-current-plan-owner';
+export { default as isECommerceTrialExpired } from './is-ecommerce-trial-expired';
 export { isRequestingSitePlans } from 'calypso/state/sites/plans/selectors/is-requesting-site-plans';
 export { isSitePlanDiscounted } from 'calypso/state/sites/plans/selectors/is-site-plan-discounted';
 export { default as isSiteOnECommerceTrial } from 'calypso/state/sites/plans/selectors/is-site-on-ecommerce-trial';
