+++ client/state/sites/plans/test/selectors.js
@@ -4,6 +4,7 @@ import {
 	PLAN_ECOMMERCE_MONTHLY,
 } from '@automattic/calypso-products';
 import deepFreeze from 'deep-freeze';
+import moment from 'moment';
 import { userState } from 'calypso/state/selectors/test/fixtures/user-state';
 import {
 	getSitePlan,
@@ -19,6 +20,9 @@ import {
 	isSitePlanDiscounted,
 	getSitePlanSlug,
 	isSiteOnECommerceTrial,
+	isECommerceTrialExpired,
+	getECommerceTrialDaysLeft,
+	getECommerceTrialExpiration,
 } from '../selectors';
 
 describe( 'selectors', () => {
@@ -784,6 +788,177 @@ describe( 'selectors', () => {
 			} );
 
 			expect( isSiteOnECommerceTrial( state, siteId ) ).toBeFalsy();
+		} );
+	} );
+
+	describe( '#getECommerceTrialExpiration()', () => {
+		const siteId = 1337;
+		test( 'Returns the expiracy date', () => {
+			const expiracyDate = '2022-02-10T00:00:00+00:00';
+
+			const plan = {
+				ID: 1,
+				productSlug: PLAN_ECOMMERCE_TRIAL_MONTHLY,
+				blogId: siteId,
+				expiryDate: expiracyDate,
+				currentPlan: true,
+			};
+
+			const state = deepFreeze( {
+				...userState,
+				sites: {
+					plans: {
+						[ siteId ]: {
+							data: [ plan ],
+						},
+					},
+					items: {
+						[ siteId ]: {
+							URL: 'https://example.wordpress.com',
+						},
+					},
+				},
+				siteSettings: {
+					items: {},
+				},
+			} );
+
+			expect(
+				getECommerceTrialExpiration( state, siteId ).isSame( moment( expiracyDate ) )
+			).toBeTruthy();
+		} );
+
+		test( 'Returns null when the trial purchase is not present', () => {
+			const plan = {};
+
+			const state = deepFreeze( {
+				...userState,
+				sites: {
+					plans: {
+						[ siteId ]: {
+							data: [ plan ],
+						},
+					},
+					items: {
+						[ siteId ]: {
+							URL: 'https://example.wordpress.com',
+						},
+					},
+				},
+				siteSettings: {
+					items: {},
+				},
+			} );
+
+			expect( getECommerceTrialExpiration( state, siteId ) ).toBeNull();
+		} );
+	} );
+
+	describe( '#getECommerceTrialDaysLeft()', () => {
+		const siteId = 1337;
+		jest.useFakeTimers().setSystemTime( new Date( '2022-01-10T00:00:00+00:00' ) );
+
+		test( 'Should return the correct number of days left before the trial expires', () => {
+			const expiracyDate = '2022-02-10T00:00:00+00:00';
+
+			const plan = {
+				ID: 1,
+				productSlug: PLAN_ECOMMERCE_TRIAL_MONTHLY,
+				blogId: siteId,
+				expiryDate: expiracyDate,
+				currentPlan: true,
+			};
+
+			const state = deepFreeze( {
+				...userState,
+				sites: {
+					plans: {
+						[ siteId ]: {
+							data: [ plan ],
+						},
+					},
+					items: {
+						[ siteId ]: {
+							URL: 'https://example.wordpress.com',
+						},
+					},
+				},
+				siteSettings: {
+					items: {},
+				},
+			} );
+
+			expect( getECommerceTrialDaysLeft( state, siteId ) ).toBe( 31 );
+		} );
+	} );
+
+	describe( '#isECommerceTrialExpired()', () => {
+		const siteId = 1337;
+		jest.useFakeTimers().setSystemTime( new Date( '2022-01-10T00:00:00+00:00' ) );
+
+		test( 'The trial period should be expired', () => {
+			const expiracyDate = '2022-01-09T00:00:00+00:00';
+
+			const plan = {
+				ID: 1,
+				productSlug: PLAN_ECOMMERCE_TRIAL_MONTHLY,
+				blogId: siteId,
+				expiryDate: expiracyDate,
+				currentPlan: true,
+			};
+
+			const state = deepFreeze( {
+				...userState,
+				sites: {
+					plans: {
+						[ siteId ]: {
+							data: [ plan ],
+						},
+					},
+					items: {
+						[ siteId ]: {
+							URL: 'https://example.wordpress.com',
+						},
+					},
+				},
+				siteSettings: {
+					items: {},
+				},
+			} );
+
+			expect( isECommerceTrialExpired( state, siteId ) ).toBeTruthy();
+		} );
+
+		test( 'The trial period should not be expired if is the same day', () => {
+			const expiracyDate = '2022-01-10T23:59:59+00:00';
+			const plan = {
+				ID: 1,
+				productSlug: PLAN_ECOMMERCE_TRIAL_MONTHLY,
+				blogId: siteId,
+				expiryDate: expiracyDate,
+				currentPlan: true,
+			};
+
+			const state = deepFreeze( {
+				...userState,
+				sites: {
+					plans: {
+						[ siteId ]: {
+							data: [ plan ],
+						},
+					},
+					items: {
+						[ siteId ]: {
+							URL: 'https://example.wordpress.com',
+						},
+					},
+				},
+				siteSettings: {
+					items: {},
+				},
+			} );
+
+			expect( isECommerceTrialExpired( state, siteId ) ).toBeFalsy();
 		} );
 	} );
 } );
